requirements.txt
streamlit
tensorflow==2.16.1
numpy
Pillow
pandas
plotly
h5py
gdown